# expression: implement vectorized evaluation for `builtinCastStringAsRealSig`