--- a/expression/bench_test.go
+++ b/expression/bench_test.go
@@ -628,6 +628,13 @@ func (g *randDurInt) gen() interface{} {
 	return int64(rand.Intn(types.TimeMaxHour)*10000 + rand.Intn(60)*100 + rand.Intn(60))
 }
 
+type randBinaryLiteralGener struct {
+}
+
+func (g *randBinaryLiteralGener) gen() interface{} {
+	return "X'01af'"
+}
+
 // locationGener is used to generate location for the built-in function GetFormat.
 type locationGener struct {
 	nullRation float64