--- a/expression/builtin_cast_vec.go
+++ b/expression/builtin_cast_vec.go
@@ -1244,6 +1244,17 @@ func (b *builtinCastStringAsRealSig) vectorized() bool {
 }
 
 func (b *builtinCastStringAsRealSig) vecEvalReal(input *chunk.Chunk, result *chunk.Column) error {
+	if IsBinaryLiteral(b.args[0]) {
+		// This block is skipped by first line of `builtinCastStringAsRealSig.evalReal`
+		// if IsBinaryLiteral(b.args[0]) {
+		// 	return b.args[0].EvalReal(b.ctx, row)
+		// }
+		if err := b.args[0].VecEvalReal(b.ctx, input, result); err != nil {
+			return err
+		}
+		return nil
+	}
+
 	n := input.NumRows()
 	bufStrings, err := b.bufAllocator.get(types.ETString, n)
 	if err != nil {